(WIP) feature: 增加日志查询配置建模(复用插件开发者中心的模型), 支持以进程维度设置查询配置
