+++ apiserver/paasng/paasng/platform/log/filters.py
@@ -16,14 +16,65 @@ limitations under the License.
 We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
-from typing import Generator, List
+import logging
+from collections import defaultdict
+from operator import attrgetter
+from typing import Counter, Dict, Generator, List, Tuple
 
+from attrs import define, field
 from elasticsearch_dsl import Search
+from rest_framework.fields import get_attribute
 
+from paasng.platform.log.constants import ESField
 from paasng.platform.modules.manager import ModuleInitializer
 from paasng.platform.modules.models import Module
+from paasng.utils.text import calculate_percentage
 
-from .constants import ESField
+logger = logging.getLogger(__name__)
+
+
+@define
+class FieldFilter:
+    # 查询字段的title
+    name: str
+    # query_term: get参数中的key
+    key: str
+    # 该field的可选项
+    options: List[Tuple[str, str]] = field(factory=list)
+    # 该 field 出现的总次数
+    total: int = 0
+
+
+def count_filters_options(logs: List, properties: Dict[str, FieldFilter]) -> List[FieldFilter]:
+    """统计 ES 日志的可选字段, 并填充到 filters 的 options"""
+    # 在内存中统计 filters 的可选值
+    field_counter: Dict[str, Counter] = defaultdict(Counter)
+    log_fields = [(f, f.split(".")) for f in properties.keys()]
+    for log in logs:
+        for log_field, split_log_field in log_fields:
+            try:
+                value = get_attribute(log, split_log_field)
+            except (AttributeError, KeyError):
+                continue
+            try:
+                field_counter[log_field][value] += 1
+            except TypeError:
+                logger.warning("Field<%s> got an unhashable value: %s", log_field, value)
+
+    result = []
+    for title, values in field_counter.items():
+        options = []
+        total = sum(values.values())
+        if total == 0:
+            # 该 field 无值可选时, 不允许使用该字段作为过滤条件
+            continue
+
+        for value, count in values.items():
+            percentage = calculate_percentage(count, total)
+            options.append((value, percentage))
+        result.append(FieldFilter(name=title, key=properties[title].key, options=options, total=total))
+    # 根据 field 在所有日志记录中出现的次数进行降序排序, 再根据 key 的字母序排序(保证前缀接近的 key 靠近在一起, 例如 json.*)
+    return sorted(result, key=attrgetter("total", "key"), reverse=True)
 
 
 class BaseAppFilter:
